add dao

+++ src/main/java/server/dao/DAO.java
@@ -16,24 +16,59 @@ public abstract class DAO<T> {
         return id;
     }
 
-    public abstract List<T> get(T pattern);
+    public List<T> get(T pattern) {
+        StringBuilder s = new StringBuilder()
+                .append("from ")
+                .append(pattern.getClass().getSimpleName());
+        String where = getWhereStatement(pattern);
+        if (0 != where.length())
+            s.append(" where ").append(where);
 
-    public abstract T getById(T pattern);
-
-    public void update(T newValue) {
         Session session = ObjectPool.getPool().getSessionFactory().openSession();
-        session.beginTransaction();
-        session.update(newValue);
-        session.getTransaction().commit();
+        List result = session.createQuery(s.toString()).list();
         session.close();
+        return result;
     }
 
-    public void delete(T value) {
+    public T getById(T pattern) {
         Session session = ObjectPool.getPool().getSessionFactory().openSession();
-        session.beginTransaction();
-        session.delete(value);
-        session.getTransaction().commit();
+        T result = (T) session.get(pattern.getClass(), getId(pattern));
         session.close();
+        return result;
     }
 
+    public boolean update(T newValue) {
+        String set = getSetStatement(newValue);
+        int id = getId(newValue);
+        if (1 > id || 0 == set.length())
+            return false;
+        StringBuilder s = new StringBuilder()
+                .append("update ")
+                .append(newValue.getClass().getSimpleName())
+                .append(" set")
+                .append(set)
+                .append(" where id=")
+                .append(id);
+
+        Session session = ObjectPool.getPool().getSessionFactory().openSession();
+        int result = session.createQuery(s.toString()).executeUpdate();
+        session.close();
+        return 1 == result;
+    }
+
+    public boolean delete(T value) {
+        if (null != getById(value)) {
+            Session session = ObjectPool.getPool().getSessionFactory().openSession();
+            session.delete(value);
+            session.close();
+            return true;
+        }
+        return false;
+    }
+
+    protected abstract String getWhereStatement(T pattern);
+
+    protected abstract String getSetStatement(T pattern);
+
+    protected abstract int getId(T value);
 }
